Adding the image to PDF converter function to kaguya_downloader.py and modifying to do in README
The image to PDF converter function first starts by looping through the list of chapter folders and then check whether the chapter already made into pdf or not. If not then the whole image inside the chapter folder are converted into pdf and stored in the download destination. This still needs more fine tuning because there might be several problem such as
1. Checking if the pdf file already exists
2. Testing the function with image that has weird numbering
3. Adding the announcement if the PDF file is created